src/evaluation/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.35.9
pydantic==2.8.0
pydantic_core==2.20.0
PyYAML==6.0.1
sniffio==1.3.1
sympy==1.12.1
torch==2.3.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
